자료형

+++ Memo/Day-4.c
@@ -2,11 +2,11 @@
 
 void main(){
     int i=10;
-    long l=10l;
+    int l=10l;
     float f=3.9f;
     double d=3.9;
     char c= 'a';
-    char * str="ABC";
+    char *str="ABC";
 
     //sizeof(자료형): 소괄호 안에 작성한 변수 또는 자료형의 크기를 알려주는 함수이다.
     printf("%d\n",sizeof(int));
